added client interface. move some api.
