--- src/oauth/oauth-storage.ts
@@ -1,11 +1,0 @@
-/*
- * Created on Thu May 14 2020
- *
- * Copyright (c) storycraft. Licensed under the MIT Licence.
- */
-
-export interface OAuthStorage {
-
-    
-
-}
